requirements.txt
Flask==3.0.0
face-recognition==1.3.0
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.1